Updated to Python3
Turns out @adosreis couldn't get the code working because he was using Python3 and I was using Python2. I've since updated to 3 and this commit makes the code work in 3. I'll let @adosreis decide if he's going to merge and convert to Python3, or ignore it and stay on Python2.